+++ frontend/src/app/app.component.ts
@@ -1,5 +1,5 @@
 import { Component } from '@angular/core';
-import { InfoPanelService } from './info-panel.service';
+import { InfoPanelService } from './services/info-panel.service';
 import { trigger, state, transition, animate, style } from '@angular/animations';
 
 @Component({
