+++ frontend/src/app/filter/filter.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { AppleService } from '../apple.service';
+import { AppleService } from '../services/apple.service';
 import { FormControl } from '@angular/forms';
 import { Apple } from '../apple';
 import { resetFakeAsyncZone } from '@angular/core/testing';
