+++ frontend/src/app/services/apple.service.ts
